refactor: Return empty list when no processors provided
- PipelineConfig has only getting BaseProcessor list from yaml configuration.
- It does not have responsibility of how creation pipeline.
