+++ source-connector/src/main/java/sourceconnector/config/PipelineConfig.java
@@ -1,13 +1,13 @@
 package sourceconnector.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import sourceconnector.domain.factory.JSONLogFactory;
+import sourceconnector.domain.log.factory.JSONLogFactory;
 import sourceconnector.domain.log.Log;
-import sourceconnector.service.processor.BaseProcessor;
-import sourceconnector.service.processor.impl.ByPassProcessor;
-import sourceconnector.service.processor.impl.EmptyFilterProcessor;
-import sourceconnector.service.processor.impl.TrimMapperProcessor;
+import sourceconnector.domain.processor.BaseProcessor;
+import sourceconnector.domain.processor.impl.EmptyFilterProcessor;
+import sourceconnector.domain.processor.impl.TrimMapperProcessor;
 
+import java.util.Collections;
 import java.util.List;
 
 @ConfigurationProperties("transform")
@@ -15,11 +15,9 @@ public record PipelineConfig(
   List<ProcessorConfig> pipeline
 ) {
 
-  // TODO: is this appropriate deciding creating ByPassProcessor when no processor is provided?
-  //  Or make ByPassProcessor list in the FileBaseLogPipeline when no list is provided
   List<BaseProcessor<Log>> toProcessors() {
     if (pipeline == null || pipeline.isEmpty()) {
-      return List.of(new ByPassProcessor());
+      return Collections.emptyList();
     }
 
     return pipeline.stream()
